Edit monitor do not monitor the build out directory
We want to monitor the edits under source tree, while the build out
directory is only used to store build outputs. During the build time,
there will be a lot of file changes under the out directory, although
the edit monitor will not send any of the file changes under the out
dir to clearcut, it needs to dispatch all the file change events
which consumes the CUP usage.

This cl will skip the build out directory when scheduling the watchdog
observer in edit monitor.

Test: atest edit_monitor_test
Bug: 381095876
Change-Id: I6759d5f7c7fd37c153bcb1fff29bd8fc37581774