aconfig: add redact param to PackageTableNode
The changes in this CL are guarded by a flag to just increment the
version to 3, as that will need a normal launch rollout. I'll make a
separate flag that we'll write to the release configs for actually
writing the redact value.

next steps for v3:
-add redact to PackageTableContext
-update cpp tests

Bug: 377311211
Flag: Build.RELEASE_ACONFIG_STORAGE_VERSION
Test: atest aconfig.test
Change-Id: I5c720d4639fed05c86895c8368ac9392090cb389