Make PRODUCT_COMPRESSED_APEX configurable in device.mk
Even though PRODUCT_COMPRESSED_APEX is a hint to use compressed APEXes,
setting it in device.mk is not respected because it's set in
updatable_apex.mk.

Now the variable is set in product_config.mk unless it's set. (default
to true to keep the behavior)

(minor fix in updatable_apex's PRODUCT_SYSTEM_PROPERTIES)

Bug: 406910790
Test: PRODUCT_COMPRESSED_APEX := false in device.mk and build
Change-Id: I935867bbc36dd809a0d3cdee2fb4bf16daff6602