Move hidl_transport soong variables to build/make
This commit is a part of changes which moves the required soong
variables from vendor/google/build to build/make, where the former is
not a PDK project, making build breakage because of missing essential
soong variables.

Conclusively, build/make is a project which can be shared with partner
projects.

Flag: EXEMPT bugfix
Bug: 407559355
Test: lunch harriet-trunk_staging-userdebug && m
Test: lunch lynx-trunk_staging-userdebug && m
Change-Id: If5a35753ce34c9bc2dc519a762b5483b199ea5bb