aconfig: add full sdk check
Update sdk check to support minor versions (using SDK_INT_FULL).

Tried a couple of different ways to represent the two types of API
levels - wanted to do an enum with values, but that wasn't parsing
into-json / from-json correctly. I debated just using u32 and then
checking if the value was > 100_000, but thought that was a bit hacky.

I went with this approach because I think it's clearer to have the
logic of finalized file -> SDK check all in the same place (here), and
then there's less logic in the template, we just place this check in
directly. Now if we have to add another check type, we don't have to
update the template.

All of the changes to the SDK_INT_FULL are behind the build flag. There
are some changes here which update the regular SDK_INT check, but they
should be very small code updates with no logic changes (as shown by the
tests) - the only actual change is to CustomFeatureFlags which is used
in testing.

Flag: Build.RELEASE_ACONFIG_SUPPORT_MINOR_SDK
Test: atest convert_finalized_flags.test
Change-Id: I05ec21aaf032eaebfbe651e4c0c2a834e3f19e18