add extended flags list for sdk 36
This change add extended flags list for sdk 36. The reason is without
this change the client code will try to read the flag value from the
storage, which makes it is hard to delete the flag in the future.

The extened flags list for 36 is computed by the difference between
enabled exported flags in the all_aconfig_declarations.pb file from
25Q2, and finalized flags in 35, 36, the extended flags for 35.

The extended flags list for 36 includes flags enabled in 25Q2, and flags
enabled in V QPR releases, 24Q4, 25Q1. The reason is these flags are
enabled in V QPR releases, they are safe to call on 25Q2 release.

Test: m android.permission.flags-aconfig-java-export and check flag
note_op_batching_enabled has sdk check
Bug: 408464400

Change-Id: Iaafe078a1a4ed8d2dafbb8edf062bdc36203c97b