Use local built libandroid_runtime.so in Robolectric.
This change builds the libaconfig_storage_read_api_cc in to the libandroid_runtime.so as a static lib when the RELEASE_READ_FROM_NEW_STORAGE flag is enabled.

I've been but a proxy for Jerome Gaillard w/ this CL.  Thank you for your deep knowledge here - I've catching up to do.

Test: CI
Bug: 399913720
Flag: HOST_ONLY
Change-Id: I9d89c243ae9e5c6a335df6d997686c5b8ae2e5cf