add skip_api_filter_list
Currently all flags in the allow_list (renamed to non_api_flags_list)
will be filtered out from the flag library for external apps.

However there is an existing flag used in SettingsLib, and need to be
shared for the usage of the apps not in Gerrit. The clean up plan for
the flag is once the feature is released, the flag can be removed from
the code without replacing to other calls. Temporary adding this file to
support this case.

Test: atest exported-flag-check-test
Bug: 408440684
Change-Id: Ic7d1c0463b5eda6ca514fa92c34dc385d7e32db9