Ensure no mixing of device config and new storage flags in a package
In a beta namespace, for a flag package, if the container is platform,
then none of the flags should use device_config, if the container is
mainline module, then none of the flags should use aconfigd. So there
should not exist a case where we can have both device_config and
aconfigd flags in a single flag package. This also simplifies codegen
template, there is no mix of old and new in a single codegen template.

Bug: b/406818480
Test: cd build/tools/aconfig/aconfig atest -c
Change-Id: Ia2cd91b95dab9b69c9a512e3077b2eaade70fe10