Remove newline from kernel version file
The kernel version file can be generated via 2 different methods:
 - Extracting it from the kernel via the extract_kernel.py script
 - Specifying it via the BOARD_KERNEL_VERSION variable, in which case
   it must still match extract_kernel.py's results.

extract_kernel.py never added a newline to the end of the file that it
generated. But in make, when using BOARD_KERNEL_VERSION, it did add
the newline.

We ended up adding a newline to the soong-built kernel version file
always, because we were comparing against a BOARD_KERNEL_VERSION
scenario. But that didn't match in the non-BOARD_KERNEL_VERSION
scenario, leading to slow builds when toggling between soong-only and
soong+make.

Bug: 412416205
Test: soong_only_diff_test.py
Change-Id: I1b6476701358c01f059ac170e2c26817f7c57c06