Allow mainline beta namespace config
When parsing flags, allow passing in a json configuration file about
mainline beta namespace. For each mainline beta namespace, the
configuration file specifies its owning container, and it exported mode
are supported for the namespace.

This configuration is used to determine the storage backend of a flag.
If it is a flag in mainline beta namespace and belongs to the owning
container, assign device config as storage backend.

Bug: b/406508083
Test: atest -c
Change-Id: I783d6c43392e102f92e46b874effbf5372cfeeee