Make jacoco-report-classes-all.jar hermetic
Like the symbols zip and proguard zips used to be, the jacoco report
zip is non-hermetic, it picks up random files dumped into a folder
as part of the rest of the build.

Apply the same resolution strategy that we've used for other packaging
zips to make the jacoco zip hermetic.

This might cause the jacoco zip to pick up fewer files than before.
We may need to add more options to pick up more test suites like
the existing JACOCO_PACKAGING_INCLUDE_DEVICE_TESTS, but it's hard to
know what's needed, so I'm going to start with this and see what issues
arise.

Bug: 413145177
Test: COVERAGE_MODULES="DOCUMENTSUI" TARGET_RELEASE=mainline_2025_06 TARGET_BUILD_VARIANT=userdebug PRODUCT=mainline_modules_x86_64 MODULE_RELEASED_PLATFORM=R ./vendor/google/build/build_unbundled_coverage_mainline_module.sh, DOCUMENTSUI-jacoco.jar didn't change.
Change-Id: I9a66d4400d9f8df3fac26be1d0fd8749d5a192f7