desktop: Pass in missing firmware dependency to pack_image
The recovery and update targets had a transitive dependncy on the
firmware targets. They were relying on the fact that the
`prebuilt-XXXX-firmware-image` was added as a `PRODUCT_PACKAGE` which
ensured the target was built beforehand. We don't actually want the
firmware.img added into the OS image since it's not used.

This change makes it so that the the recovery and update images directly
depend on the `gen-XXXX-firmware-image` genrule. This will allow us to
remove the `prebuilt-XXXX-firmware-image` `prebuilt_etc` target and drop
the hard coded path to the outputs.

In a followup CL I will merge all the `gen-XXXX-firmware-images` into a
single `desktop-firmware-squashfs` genrule. I will do this once all
devices are using the firmware shellball.

Output of pack_image:
```
>> Writing out/target/product/brya/obj/ETC/gen-brya-firmware-image_intermediates/gen-brya-firmware-image to out/target/product/brya/android-desktop_update_image.bin in partition 5
+ dd if=out/target/product/brya/obj/ETC/gen-brya-firmware-image_intermediates/gen-brya-firmware-image of=out/target/product/brya/android-desktop_update_image.bin oflag=seek_bytes bs=4M seek=9336520704 conv=notrunc
64+1 records in
64+1 records out
```

Bug: 419155941
Test: m installclean && m out/target/product/brya/android-desktop_recovery_image.bin
Flag: EXEMPT desktop only
Change-Id: Ib56f14403d0fb712d03166f15992572d73f65567