finalization-test: correctly parse bool flag values
When reading the protobuf data, finalization-test converts all flag
values into Strings (for simplicity). It incorrectly assumed that a
flag's `string_value()` would always return a string representation of
the value, but this is only true if the underlying type is a string.

Update the logic to explicitly convert strings, bools and empty values
to Strings.

Test: atest finalization-test
Bug: 420551526
Change-Id: I5fde8e70f490fabd6b11575dcf6d0b90644eea7e