Reapply "Build apkcerts.txt with soong"
This reverts commit e4e5103a590704e4688109fc9cf9631cd129f807.

Make will still handle building apkcerts.txt for make modules, but
soong will build it for soong modules and then make will combine
the make and soong apkcerts files.

This is for soong-only builds. In a followup change I'll make it so
that only soong dists the apkcerts.txt file on unbundled builds.

The difference in this reapplication from the origional cl is the
addition of soong_apkcerts_file_with_soong_and_make_modules_removed,
as make was seeing some modules that are in both make and soong as
make modules, and not including them in the soong doublecheck file.

Bug: 413145177
Test: Presubmits, make will doublecheck that the soong-built apkcerts.txt is correct
Change-Id: I4bf6304503951aa88c3d2573ffe99d24e54ae169