Add filter and replace flags to textnotice
When soong builds text notice files, it does so for all possible files
that the module outputs, even though we might not be using all of them.
Add some filter flags that can be used to restrict the notice file
to only the files that we care about.

Also add a replace flag for path substitutions. This is to support
the equivalent of the "root mappings" in declare-container-license-deps
in make.

Bug: 422276561
Test: Manually
Change-Id: Ib37571b420b56231a3884d120f61ff5ebe68e29f