Minimal fix for fixing Asan LeakSanitizer error
This is the best-effort fix for the LeakSanitizer error of reporting
leak of a flag_provider object.

The restriction is that `provider_` is `extern` and not `const` either.
External libraries mock the flag_provider_interface and injects an
instance. We can't easily make the actual flag_provider instance
function-local static. In this CL, the instance is still malloc-ed.

Flag: EXEMPT bugfix
Bug: 356201402
Test: atest aconfig.test
Test: m test-art-host-gtest-art_compiler_tests64 under asan build
Change-Id: I6330a6e5dc498924dee31571fadd0fec1d1620be