Suppress error logging when storage file is not available on host
The files that store flag value overrides on Android devices are absent
from host devices. As a result, the default values are always returned
when running on the host. This change prevents error messages caused by
the absence of storage files when running on the host.

Bug: 419824590

Test: atest aconfig.test

Flag: EXEMPT host code

Change-Id: I4e6185284e0726846c5cf36ac036d18fd1d51262