Default journal_size and extfs_rsv_pct for user specified mount point
build_image.py implicitly adds some properties on top of the properties
in the user provided prop file. Some of these were added only
when mount_point was not explicitly specified. This was causing a
difference in *.img created by Make and Soong. The latter explicitly
specificies `mount_point` in the prop file.

To fix these issues, the implicit addition has been copied over to the
main parser.

Bug: 419860251
Test: lunch husky-trunk_staging-userdebug
Test: m out/target/product/husky/system_ext.img out/soong/.intermediates/build/soong/fsgen/husky_generated_system_ext_image/android_common/system_ext.img
Test: verified that both are bit-identical
Change-Id: I4167cfc813be94fdf80ec9c830a1aa63bbe56f9f