Remove appcompat.zip
I confirmed with the origional author, dariofreni@, that this is no
longer used.

Also, it's been broken and always producing an empty zip file since
aosp/3106579, where a trailing backslash caused the `else` to not be
a functional makefile if-else, so the else block was always taken,
and the appcompat functions were reset to empty strings after defining
them.

Fixes: 428050724
Test: Presubmits
Change-Id: I12d67a0a4c7b2442040867d3629311310e9a44ae