Add a soong variable to conditionally allow profile code
Low-overhead tracing is an experimental feature that is guarded behind a
read-write flag and is targeted only for pixel devices. To have fine
granular control to allow this feature we add a soong variable and set
it to true only for pixel devices.

Bug: 352518093
Test: art/test.py
Flag: com.android.art.rw.flags.enable_profile_code_rw

Change-Id: I8b5b6bcdad639ad7619ed9a241ec1053a455e167