Change the path of the Android output directory on creation when Cartfs is running.
When Cartfs is running, the output directory will be created under /google/cartfs/mount/<worksapce name>/out.

Additionally, lookup existing folders in cartfs and verify if the workspace still exists. If not, delete the folder in cartfs. This is needed until we have a better solution that will work for everyone.

This is needed as part of the CoG acceleration plan, which uses Cartfs as the output directory to resolve some of the issues.

Relevant docs:
* go/cog-acceleration
* go/cartfs
* go/bescet-artfs-as-cog-symlink-target
* Launch plan: go/buildhive-artfs-cog-out-dir-launch-plan

BUG: 418985556
TESTED=Tested on cloud top with Cartfs running and not running.

Change-Id: I3d761660ea15eb355e9e75ca5996f593b27f8dd0