Move prebuilt radio image logic to build/make
The packaging rule for the prebuilt radio image file is currently
defined in the vendor mk files, by calling the `add-radio-file-checked`
macro.

This change moves the logic out of the vendor mk file, and conditionally
defines the packaging rule based on the definition of
`BOARD_RADIO_IMAGE_PATH` variable.

Note that the rule is defined in `android-info.mk` instead of the core
Makefile. This is because the `CHECK_RADIO_VERSIONS`, which depends on
`BOARD_INFO_CHECK`, is run before the Makefile is include in the top
level main.mk file.

Test: build/soong/scripts/soong_only_diff_test.py factory_tegu
Bug: 432817167
Bug: 434279647
Change-Id: I493e8937327b0c475b5c96fcb968cfc5fdb84a21