Install EyeDropper on devices derived from handheld_system
This package is used to provide a new public API
android.content.Intent.OPEN_EYE_DROPPER. That gives apps a way to
request the user to select a pixel from displays, and return the color
of the selected pixel to the requesting apps.

This is more useful in desktop mode, but it also provides similar
functionality on phones and tablets out of desktop mode, so we should
install it on them too, similarly to DocumentsUI.

The API is very unlikely to be used in system user in HSUM, so there is
no need to have it installed there.

Bug: 399172476
Test: Builds
Flag: RELEASE_ACONFIG_INCLUDE_EYE_DROPPER_APP
Change-Id: I705a2a8003b99c1d5f3499879cbf250654ae2ef6