Abstract get_file() from CheckCombinedSepolicy
This CL doesn't affect the original CheckCombinedSepolicy behavior. But
makes it flexible for users to decide the ways how these files are
retrieved.

For example, when the input is not target files dir but target files
zip, a custom get_file() implementation can be passed in to extract
files from the zip.

Bug: 429283294
Test: presubmit
Change-Id: Iea7e6f03e2bc78a35bd3984de98cee7d052d148f