releasetools: Ensure pvmfw.img is always copied
The signing script would only write pvmfw.img to the output if its
embedded public key was being replaced. If no key replacement occurred,
the file was dropped, causing a downstream assertion failure.

This change ensures that if the key is not updated, the original
pvmfw.img is copied verbatim to the output target files.

Bug: 438711111
Test: m sign_target_files_apks
Test: sign_target_files_apks --allow_gsi_debug_sepolicy \
         --extra_apex_payload_key com.android.virt.apex= \
         -e com.android.virt.apex= \
         gsi_car_arm64-target_files-${build_id}.zip signed.zip
Change-Id: Ifb0bf393be8064e1c0b123f4b9d21c20c0338355