desktop: Replace unsigned FW with presigned FW during signing
The Android signer doesn't poses the ability to sign desktop AP/EC
firmware. To workaround this we commit pre-signed firmware into the
`<product>-prebuilts/firmware/{dev,premp}-signed` directories.

The `desktop-firmware-zip-signed` target will be generating a `zip` file
with the signed firmware. This firmware should only ever be used with a
signed OS image. We never want to mix unsigned firmware with a signed OS
since it won't boot, or vice versa.

When a desktop build is built, it generates a
`android-desktop_image.bin`. This file is "unsigned" since the builders
don't have access to real keys. For this reason we cant just swap out
the firmware during the build.

This change solves the issue by adding the signed firmware into the
`target_files.zip`. The `sign_target_files_apks.py` (which is only
called by the signer after the builder has completed) then swaps out the
unsigned firmware with the signed firmware. This solves the OTA case
where we bundle the signed firmware with the signed OS.

Bug: 435006163
Flag: EXEMPT desktop only change
Test: m target-files-package
Test: ./out/host/linux-x86/bin/sign_target_files_apks out/target/product/brya/obj/PACKAGING/target_files_intermediates/brya-target_files.zip ~/tmp/signed-target_files.zip
Change-Id: Ic1bf77da520b328f827e580c860ee998ca4a5c18